platform.rs -> application.rs before merge (any way to keep the rename without commiting a revert before every merge?)
